refactor: divide app and components
