+++ src/app/intro/ui/Banner/AdSlide.tsx
