+++ src/app/intro/ui/Banner/AdSlider.tsx
