+++ src/app/intro/ui/Banner/Container.tsx
