+++ src/app/intro/ui/Banner/Job.tsx
