+++ src/app/intro/ui/Examples/BannerCard.tsx
