+++ src/app/intro/ui/Examples/JobCard.tsx
